@@ -7,7 +7,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.5.3
+  sdk: ^3.6.0
 
 dependencies:
   cached_network_image: ^3.4.1
@@ -21,11 +21,11 @@ dependencies:
   flutter_dotenv: ^5.2.1
   flutter_native_splash: ^2.4.3
   flutter_secure_storage: ^9.2.2
-  flutter_svg: ^2.0.15
+  flutter_svg: ^2.0.16
   font_awesome_flutter: ^10.8.0
-  get_it: ^8.0.2
-  lottie: ^3.1.3
-  go_router: ^14.6.1
+  get_it: ^8.0.3
+  lottie: ^3.3.0
+  go_router: ^14.6.2
   google_sign_in: ^6.2.2
   icons_launcher: ^3.0.0
   image_picker: ^1.1.2
@@ -39,7 +39,6 @@ dependencies:
   country_code_picker: ^3.1.0
   pinput: ^5.0.0
   google_generative_ai: ^0.4.6
-  flutter_gemini: ^2.0.5
 
 dev_dependencies:
   build_runner: ^2.4.13
@@ -47,6 +46,12 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
   objectbox_generator: ^4.0.3
+
+#dependency_overrides:
+#  file_picker:
+#    git:
+#      url: https://github.com/abdelaziz-mahdy/flutter_file_picker
+#      ref: fix-flutter-3.27-compile-error
 
 flutter:
   uses-material-design: true
