@@ -27,7 +27,8 @@ class LoadingDialogManager {
       builder: (BuildContext context) {
         return Dialog(
           insetPadding: EdgeInsets.symmetric(horizontal: 100),
-          backgroundColor: context.theme.appColors.background.withOpacity(0.9),
+          backgroundColor:
+              context.theme.appColors.background.withValues(alpha: 0.9),
           shape:
               RoundedRectangleBorder(borderRadius: BorderRadius.circular(20)),
           child: ConstrainedBox(
